Resolving duplicate types with common header
